Extract tests from PrintReadsIntegrationTest to share with Spark version.
